backend/requirements.txt
gunicorn==20.0.4
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
six==1.15.0
Werkzeug==1.0.1
